requirements.txt
pillow==10.4.0
colorgram.py==1.2.0
Flask==3.0.3